refactor: service 추상화 제거
